pyproject.toml
[project]
name = "spyder-to-get-number"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openpyxl>=3.1.5",
    "pandas>=2.2.3",
    "pymysql>=1.1.1",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
    "sqlalchemy>=2.0.38",
    "tqdm>=4.67.1",
    "urllib3==1.26.5",
]
